chore: reorganization of modules
