+++ common/api-traits/src/chain/channels.rs
@@ -1,0 +1,64 @@
+use std::future::Future;
+
+use futures::stream::BoxStream;
+pub use hopr_internal_types::prelude::{ChannelDirection, ChannelEntry, ChannelId, ChannelStatus};
+use hopr_primitive_types::prelude::Address;
+pub use hopr_primitive_types::prelude::HoprBalance;
+
+use crate::chain::ChainReceipt;
+
+/// Selector for channels.
+///
+/// See [`ChainReadChannelOperations::stream_channels`].
+#[derive(Debug, Clone, PartialEq, Eq, Default)]
+pub struct ChannelSelector {
+    /// Filter by counterparty address.
+    pub counterparty: Option<Address>,
+    /// Filter by direction.
+    pub direction: Option<ChannelDirection>,
+    /// Filter by possible channel states.
+    pub allowed_states: Vec<ChannelStatus>,
+}
+
+/// On-chain read operations regarding channels.
+#[async_trait::async_trait]
+pub trait ChainReadChannelOperations {
+    type Error;
+
+    /// Returns a single channel given `src` and `dst`.
+    async fn channel_by_parties(&self, src: &Address, dst: &Address) -> Result<Option<ChannelEntry>, Self::Error>;
+
+    /// Returns a single channel given `channel_id`.
+    async fn channel_by_id(&self, channel_id: &ChannelId) -> Result<Option<ChannelEntry>, Self::Error>;
+
+    /// Returns a stream of channels given the [`ChannelSelector`].
+    async fn stream_channels<'a>(
+        &'a self,
+        selector: ChannelSelector,
+    ) -> Result<BoxStream<'a, Result<ChannelEntry, Self::Error>>, Self::Error>;
+}
+
+/// On-chain write operations regarding channels.
+#[async_trait::async_trait]
+pub trait ChainWriteChannelOperations {
+    type Error;
+    /// Opens a channel with `dst` and `amount`.
+    async fn open_channel(
+        &self,
+        dst: &Address,
+        amount: HoprBalance,
+    ) -> Result<impl Future<Output = Result<ChainReceipt, Self::Error>>, Self::Error>;
+
+    /// Funds an existing channel.
+    async fn fund_channel(
+        &self,
+        channel_id: &ChannelId,
+        amount: HoprBalance,
+    ) -> Result<impl Future<Output = Result<ChainReceipt, Self::Error>>, Self::Error>;
+
+    /// Closes an existing channel.
+    async fn close_channel(
+        &self,
+        channel_id: &ChannelId,
+    ) -> Result<impl Future<Output = Result<ChainReceipt, Self::Error>>, Self::Error>;
+}
