+++ common/api-traits/src/chain/mod.rs
@@ -1,0 +1,12 @@
+mod accounts;
+mod channels;
+mod keys;
+mod tickets;
+
+/// Receipt of an on-chain operation.
+pub type ChainReceipt = hopr_crypto_types::prelude::Hash;
+
+pub use accounts::*;
+pub use channels::*;
+pub use keys::*;
+pub use tickets::*;
