+++ common/api-traits/src/lib.rs
@@ -1,0 +1,5 @@
+/// On-chain operations-related API traits.
+pub mod chain;
+
+pub use hopr_crypto_types::prelude::OffchainPublicKey;
+pub use hopr_primitive_types::prelude::Address;
