--- common/traits/src/lib.rs
@@ -1,3 +1,0 @@
-mod chain;
-
-pub use chain::*;
